Separate questDetails
