+++ client/src/hooks/helpers/useQuests.tsx
@@ -8,28 +8,7 @@ import { useEntities } from "./useEntities";
 import { ArmyInfo, useArmiesByEntityOwner } from "./useArmies";
 import { useGetMyOffers } from "./useTrade";
 import { getPillageEvents } from "@/dojo/events/pillageEventQueries";
-
-export enum QuestName {
-  Settle = "Settle",
-  BuildFarm = "Build a Farm",
-  BuildResource = "Build a Resource Facility",
-  CreateTrade = "Create a Trade",
-  CreateArmy = "Create an Army",
-  Travel = "Travel with your Army",
-  BuildWorkersHut = "Build a workers hut",
-  Market = "Build a market",
-  Pillage = "Pillage a structure",
-  Mine = "Discover an earthenshard mine",
-  Contribution = "Contribute to a hyperstructure",
-  Hyperstructure = "Build a hyperstructure",
-}
-
-interface StaticQuestInfo {
-  description: string;
-  steps: string[];
-  prizes: Prize[];
-  depth: number;
-}
+import { QuestName, questDetails } from "@/ui/components/quest/QuestDetails";
 
 export interface Quest {
   name: QuestName;
@@ -50,132 +29,6 @@ export enum QuestStatus {
   Completed,
   Claimed,
 }
-
-const questInfo = new Map<QuestName, StaticQuestInfo>([
-  [
-    QuestName.Settle,
-    {
-      description: "A gift of food from the gods",
-      steps: ["Settle your first Realm"],
-      prizes: [{ id: QuestType.Food, title: "Common Resources" }],
-      depth: 0,
-    },
-  ],
-  [
-    QuestName.BuildFarm,
-    {
-      description: "Wheat is the lifeblood of your people. Go to the construction menu and build a farm",
-      steps: [
-        "Navigate to the construction menu",
-        "Select the 'Farm' card",
-        "Left click on a hex to build it, or right click to cancel",
-      ],
-      prizes: [
-        { id: QuestType.CommonResources, title: "Common Resources" },
-        { id: QuestType.UncommonResources, title: "Uncommon Resources" },
-        { id: QuestType.RareResources, title: "Rare Resources" },
-        { id: QuestType.UniqueResources, title: "Unique Resources" },
-        { id: QuestType.LegendaryResources, title: "Legendary Resources" },
-        { id: QuestType.MythicResources, title: "Mythic Resources" },
-      ],
-      depth: 1,
-    },
-  ],
-  [
-    QuestName.BuildResource,
-    {
-      description: "Eternum thrives on resources. Construct resource facilities to harvest them efficiently",
-      steps: [
-        "Navigate to the construction menu",
-        "Navigate to the 'Resources' tab and select a resource card",
-        "Left click on a hex to build it, or right click to cancel",
-      ],
-      prizes: [{ id: QuestType.Trade, title: "Donkeys and Lords" }],
-      depth: 2,
-    },
-  ],
-  [
-    QuestName.CreateTrade,
-    {
-      description: "Trading is the lifeblood of Eternum. Create a trade to start your economy",
-      steps: [],
-      prizes: [{ id: QuestType.Military, title: "Claim Starting Army" }],
-      depth: 3,
-    },
-  ],
-  [
-    QuestName.CreateArmy,
-    {
-      description: "Conquest is fulfilling. Create an army to conquer your enemies",
-      steps: ["Create an army to conquer your enemies", "Assign troops to your army"],
-      prizes: [{ id: QuestType.Earthenshard, title: "Claim Earthen Shard" }],
-      depth: 4,
-    },
-  ],
-  [
-    QuestName.Travel,
-    {
-      description: "Travel with your army",
-      steps: ["Go to world view", "Right click on your army", "Travel w/ your army"],
-      prizes: [{ id: QuestType.Travel, title: "Travel" }],
-      depth: 5,
-    },
-  ],
-  [
-    QuestName.BuildWorkersHut,
-    {
-      description: "Build worker huts to extend your population capacity",
-      steps: [],
-      prizes: [{ id: QuestType.Population, title: "Population" }],
-      depth: 6,
-    },
-  ],
-  [
-    QuestName.Market,
-    {
-      description: "Build a market to produce donkeys. Donkeys are a resource used to transport goods",
-      steps: [],
-      prizes: [{ id: QuestType.Market, title: "Market" }],
-      depth: 6,
-    },
-  ],
-  [
-    QuestName.Pillage,
-    {
-      description: "Pillage a realm, hyperstructure or earthenshard mine",
-      steps: [],
-      prizes: [{ id: QuestType.Pillage, title: "Pillage" }],
-      depth: 6,
-    },
-  ],
-  [
-    QuestName.Mine,
-    {
-      description: "Explore the world, find earthenshard mines",
-      steps: [],
-      prizes: [{ id: QuestType.Mine, title: "Mine" }],
-      depth: 6,
-    },
-  ],
-  [
-    QuestName.Contribution,
-    {
-      description: "Contribute to a Hyperstructure",
-      steps: [],
-      prizes: [{ id: QuestType.Contribution, title: "Contribution" }],
-      depth: 6,
-    },
-  ],
-  [
-    QuestName.Hyperstructure,
-    {
-      description: "Build a Hyperstructure",
-      steps: [],
-      prizes: [{ id: QuestType.Hyperstructure, title: "Hyperstructure" }],
-      depth: 6,
-    },
-  ],
-]);
 
 export const useQuests = () => {
   const {
@@ -331,7 +184,7 @@ const useBuildQuests = () => {
   const useSettleQuest = (): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Settle)!,
+        ...questDetails.get(QuestName.Settle)!,
         name: QuestName.Settle,
         status: questClaimStatus[QuestName.Settle] ? QuestStatus.Claimed : QuestStatus.Completed,
       }),
@@ -342,7 +195,7 @@ const useBuildQuests = () => {
   const useBuildFarmQuest = (farmCount: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.BuildFarm)!,
+        ...questDetails.get(QuestName.BuildFarm)!,
         name: QuestName.BuildFarm,
         status:
           farmCount > 0
@@ -358,7 +211,7 @@ const useBuildQuests = () => {
   const useBuildResourceQuest = (resourceCount: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.BuildResource)!,
+        ...questDetails.get(QuestName.BuildResource)!,
         name: QuestName.BuildResource,
         status:
           resourceCount > 0
@@ -374,7 +227,7 @@ const useBuildQuests = () => {
   const useCreateTradeQuest = (orders: any[]): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.CreateTrade)!,
+        ...questDetails.get(QuestName.CreateTrade)!,
         name: QuestName.CreateTrade,
         status:
           orders.length > 0
@@ -390,7 +243,7 @@ const useBuildQuests = () => {
   const useCreateArmyQuest = (entityArmies: ArmyInfo[], hasTroops: boolean): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.CreateArmy)!,
+        ...questDetails.get(QuestName.CreateArmy)!,
         name: QuestName.CreateArmy,
         status:
           entityArmies.length > 0 && hasTroops
@@ -406,7 +259,7 @@ const useBuildQuests = () => {
   const useTravelQuest = (hasTraveled: boolean): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Travel)!,
+        ...questDetails.get(QuestName.Travel)!,
         name: QuestName.Travel,
         status: hasTraveled
           ? questClaimStatus[QuestName.Travel]
@@ -421,7 +274,7 @@ const useBuildQuests = () => {
   const useBuildWorkersHutQuest = (workersHutCount: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.BuildWorkersHut)!,
+        ...questDetails.get(QuestName.BuildWorkersHut)!,
         name: QuestName.BuildWorkersHut,
         status:
           workersHutCount > 0
@@ -437,7 +290,7 @@ const useBuildQuests = () => {
   const useMarketQuest = (marketsCount: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Market)!,
+        ...questDetails.get(QuestName.Market)!,
         name: QuestName.Market,
         status:
           marketsCount > 0
@@ -453,7 +306,7 @@ const useBuildQuests = () => {
   const usePillageQuest = (pillageHistoryLength: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Pillage)!,
+        ...questDetails.get(QuestName.Pillage)!,
         name: QuestName.Pillage,
         status:
           pillageHistoryLength > 0
@@ -469,7 +322,7 @@ const useBuildQuests = () => {
   const useMineQuest = (fragmentMines: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Mine)!,
+        ...questDetails.get(QuestName.Mine)!,
         name: QuestName.Mine,
         status:
           fragmentMines > 0
@@ -485,7 +338,7 @@ const useBuildQuests = () => {
   const useContributionQuest = (hyperstructureContributions: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Contribution)!,
+        ...questDetails.get(QuestName.Contribution)!,
         name: QuestName.Contribution,
         status:
           hyperstructureContributions > 0
@@ -501,7 +354,7 @@ const useBuildQuests = () => {
   const useHyperstructureQuest = (hyperstructures: number): Quest => {
     return useMemo(
       () => ({
-        ...questInfo.get(QuestName.Hyperstructure)!,
+        ...questDetails.get(QuestName.Hyperstructure)!,
         name: QuestName.Hyperstructure,
         status:
           hyperstructures > 0
@@ -556,10 +409,10 @@ export const useQuestClaimStatus = () => {
   };
 
   const questClaimStatus = useMemo(() => {
-    return Array.from(questInfo.keys()).reduce(
+    return Array.from(questDetails.keys()).reduce(
       (acc, questName) => ({
         ...acc,
-        [questName]: checkPrizesClaimed(questInfo.get(questName)?.prizes || []),
+        [questName]: checkPrizesClaimed(questDetails.get(questName)?.prizes || []),
       }),
       {} as Record<QuestName, boolean>,
     );
